Fix unit failed unit tests due to reorganize file structure
