+++ lib/__tests__/DatePicker/Calendar.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow, utilsToUse } from '../test-utils';
-import { Calendar } from '../../src/DatePicker/Calendar';
+import { Calendar } from '../../src/DatePicker/components/Calendar';
 
 describe('Calendar', () => {
   let component;
