+++ lib/__tests__/DatePicker/CalendarHeader.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow, utilsToUse } from '../test-utils';
-import { CalendarHeader } from '../../src/DatePicker/CalendarHeader';
+import { CalendarHeader } from '../../src/DatePicker/components/CalendarHeader';
 
 describe('CalendarHeader', () => {
   let component;
