+++ lib/__tests__/DatePicker/DatePickerModal.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow, utilsToUse } from '../test-utils';
-import { DatePickerWrapper } from '../../src/DatePicker/DatePickerWrapper';
+import { DatePickerModal } from '../../src/DatePicker/DatePickerModal';
 
 const spy = jest.fn();
 
@@ -11,11 +11,11 @@ const props = {
   value: utilsToUse.date('2018', 'YYYY'),
 };
 
-describe('DatePickerWrapper', () => {
+describe('DatePickerModal', () => {
   let component;
 
   beforeEach(() => {
-    component = shallow(<DatePickerWrapper {...props} />);
+    component = shallow(<DatePickerModal {...props} />);
   });
 
   it('Should renders', () => {
