+++ lib/__tests__/DatePicker/YearSelection.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow, utilsToUse } from '../test-utils';
-import { YearSelection } from '../../src/DatePicker/YearSelection';
+import { YearSelection } from '../../src/DatePicker/components/YearSelection';
 
 describe('YearSelection', () => {
   let component;
