+++ lib/__tests__/DateTimePicker/DateTimePickerHeader.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow, utilsToUse } from '../test-utils';
-import { DateTimePickerHeader } from '../../src/DateTimePicker/DateTimePickerHeader';
+import { DateTimePickerHeader } from '../../src/DateTimePicker/components/DateTimePickerHeader';
 
 describe('DateTimePickerHeader', () => {
   let component;
