+++ lib/__tests__/DateTimePicker/DateTimePickerModal.test.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { shallow } from '../test-utils';
-import { DateTimePickerWrapper } from '../../src/DateTimePicker/DateTimePickerWrapper';
+import { DateTimePickerModal } from '../../src/DateTimePicker/DateTimePickerModal';
 
-describe('DateTimePickerWrapper', () => {
+describe('DateTimePickerModal', () => {
   let component;
 
   beforeEach(() => {
-    component = shallow(<DateTimePickerWrapper classes={{}} />);
+    component = shallow(<DateTimePickerModal classes={{}} />);
   });
 
   it('Should renders', () => {
