+++ lib/__tests__/DateTimePicker/DateTimePickerTabs.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from '../test-utils';
-import { DateTimePickerTabs } from '../../src/DateTimePicker/DateTimePickerTabs';
+import { DateTimePickerTabs } from '../../src/DateTimePicker/components/DateTimePickerTabs';
 
 describe('DateTimePickerTabs', () => {
   let component;
