+++ lib/__tests__/DateTimePicker/DateTimePickerView.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from '../test-utils';
-import { DateTimePickerView } from '../../src/DateTimePicker/DateTimePickerView';
+import { DateTimePickerView } from '../../src/DateTimePicker/components/DateTimePickerView';
 
 describe('DateTimePickerView', () => {
   let component;
