+++ lib/__tests__/TimePicker/Clock.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from '../test-utils';
-import { Clock } from '../../src/TimePicker/Clock';
+import { Clock } from '../../src/TimePicker/components/Clock';
 
 describe('Clock', () => {
   let component;
