+++ lib/__tests__/TimePicker/ClockNumber.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from '../test-utils';
-import { ClockNumber } from '../../src/TimePicker/ClockNumber';
+import { ClockNumber } from '../../src/TimePicker/components/ClockNumber';
 
 describe('ClockNumber', () => {
   let component;
