+++ lib/__tests__/TimePicker/ClockPointer.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from '../test-utils';
-import { ClockPointer } from '../../src/TimePicker/ClockPointer';
+import { ClockPointer } from '../../src/TimePicker/components/ClockPointer';
 
 describe('ClockPointer', () => {
   let component;
