+++ lib/__tests__/TimePicker/TimePickerModal.test.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { shallow } from '../test-utils';
-import { TimePickerWrapper } from '../../src/TimePicker/TimePickerWrapper';
+import { TimePickerModal } from '../../src/TimePicker/TimePickerModal';
 
-describe('TimePickerWrapper', () => {
+describe('TimePickerModal', () => {
   let component;
 
   beforeEach(() => {
-    component = shallow(<TimePickerWrapper />);
+    component = shallow(<TimePickerModal />);
   });
 
   it('Should renders', () => {
