+++ lib/__tests__/TimePicker/TimePickerView.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow, utilsToUse } from '../test-utils';
-import { TimePickerView } from '../../src/TimePicker/TimePickerView';
+import { TimePickerView } from '../../src/TimePicker/components/TimePickerView';
 
 describe('TimePickerView', () => {
   let component;
