+++ lib/__tests__/e2e/DateTimePickerModal.test.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { mount, utilsToUse } from '../test-utils';
-import { DateTimePickerWrapper } from '../../src/DateTimePicker/DateTimePickerWrapper';
+import { DateTimePickerModal } from '../../src/DateTimePicker/DateTimePickerModal';
 
 const format = process.env.UTILS === 'luxon' ? 'MM/dd/yyyy hh:mm' : 'MM/DD/YYYY HH:mm';
-describe('e2e - DateTimePickerWrapper', () => {
+describe('e2e - DateTimePickerModal', () => {
   let component;
   const onChangeMock = jest.fn();
 
   beforeEach(() => {
-    component = mount(<DateTimePickerWrapper
+    component = mount(<DateTimePickerModal
       classes={{}}
       format={format}
       value={utilsToUse.date('2018-01-01T00:00:00.000Z')}
@@ -53,13 +53,13 @@ describe('e2e - DateTimePickerWrapper', () => {
 });
 
 
-describe('e2e - DateTimePickerWrapper keyboard', () => {
+describe('e2e - DateTimePickerModal keyboard', () => {
   let component;
   const onChangeMock = jest.fn();
   const onInputChangeMock = jest.fn();
 
   beforeEach(() => {
-    component = mount(<DateTimePickerWrapper
+    component = mount(<DateTimePickerModal
       keyboard
       classes={{}}
       format={format}
